Explicitly disable timezone info in SQLAlchemy model
SQLAlchemy can run into issues when storing datetime values in MySQL due
to the latter missing support for timezone details in the DATETIME type.
One way to trigger this is to run Gryphon on a server with a non-UTC
system timezone. The workaround is to start over with a fresh setup and
UTC system timezone, or disable timezone info in the SQLAlchemy model.

Example exception triggered by this issue:
  ERROR:gryphon.execution.live_runner:[OperationalError]
  (_mysql_exceptions.OperationalError) (1292, "Incorrect datetime value:
  '2019-07-12 15:29:13+00:00' for column 'time_created' at row 1")